+++ esp-radio/src/wifi/mod.rs
@@ -124,6 +124,7 @@ use crate::binary::{
         esp_wifi_set_mode,
         esp_wifi_set_protocol,
         esp_wifi_set_tx_done_cb,
+        esp_wifi_sta_get_ap_info,
         esp_wifi_sta_get_rssi,
         esp_wifi_start,
         esp_wifi_stop,
@@ -3130,6 +3131,31 @@ impl WifiController<'_> {
             // Will return ESP_FAIL -1 if called in AP mode.
             esp_wifi_result!(unsafe { esp_wifi_sta_get_rssi(&mut rssi) })?;
             Ok(rssi)
+        } else {
+            Err(WifiError::Unsupported)
+        }
+    }
+
+    /// Get the Access Point information of AP to which the device is associated with.
+    /// The value is obtained from the last beacon.
+    ///
+    /// <div class="warning">
+    ///
+    /// - Use this API only in STA or AP-STA mode.
+    /// - This API should be called after the station has connected to an access point.
+    /// </div>
+    ///
+    /// # Errors
+    /// This function returns [`WifiError::Unsupported`] if the STA side isn't
+    /// running. For example, when configured for AP only.
+    pub fn ap_info(&self) -> Result<AccessPointInfo, WifiError> {
+        if self.mode()?.is_sta() {
+            let mut record: MaybeUninit<include::wifi_ap_record_t> = MaybeUninit::uninit();
+            esp_wifi_result!(unsafe { esp_wifi_sta_get_ap_info(record.as_mut_ptr()) })?;
+
+            let record = unsafe { MaybeUninit::assume_init(record) };
+            let ap_info = convert_ap_info(&record);
+            Ok(ap_info)
         } else {
             Err(WifiError::Unsupported)
         }